Comply with ITU-R BT.500-13 recommendation
This commit performs changes to comply with the recommendation
ITU-R BT.500-13 "Methodology for the subjective assessment of
the quality of television pictures".

The presentation order of each pair is now fixed in the
(reference, test) order instead of the random order previously used.

The stimuli are now presented just once instead of twice, with a
longer view duration per image.